+++ docs/CHANGELOG.md
@@ -10,7 +10,8 @@ The format of this change log follows the advice given at [Keep a CHANGELOG](htt
 
 ## [Unreleased]
 ### Added
-- Added support for the `--tags` and `--name` options into the `behat` command.
+- Added support for the `--tags` and `--name` options to the `behat` command.
+- Added support for the `--configure`, `--testsuite` and `--filter` options to the `phpunit` command.
 
 ### Changed
 - ACTION SUGGESTED: If you are using GitHub Actions, it's recomended to use `!cancelled()` instead of `always()` for moodle-plugin-ci tests. Adding a final step that always returns failure when the workflow is cancelled will ensure that cancelled workflows are not marked as successful. For a working example, please reference the updated `gha.dist.yml` file.